+++ .gitignore
@@ -3,3 +3,5 @@ __pycache__/semantic_search.cpython-312.pyc
 .env
 .env
 /__pycache__
+/services/__pycache__
+/utils/__pycache__
