--- static/css/custom-design.css
@@ -1,7 +1,0 @@
-.card, .list-group-item {
-    border-radius: 8px;
-    box-shadow: 0 1px 4px rgba(0,0,0,0.15);
-}
-button {
-    transition: background-color 0.2s, border-color 0.2s;
-}
